api/django-app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
billiard==4.2.2
bleach==6.2.0
boto3==1.40.59
botocore==1.40.59
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
diff-match-patch==20241021
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-import-export==4.3.10
django-ninja==1.4.5
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.1
django-unfold==0.68.0
django_celery_results==2.6.0
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
kombu==5.5.4
Markdown==3.9
martor==1.7.15
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.0.0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.11.0
zipp==3.23.0
